BTFS-652: Unit tests for challenge gen & solve
Moved part of coreunix's add tests to a common add.go so that both
add_test and challenge_test can share the same reed solomon based file
addition routine.

Added a gen & solve flow test for 100 questions & answers.
