+++ core/coreunix/test/add_test.go
@@ -3,8 +3,6 @@ package test
 import (
 	"bytes"
 	"context"
-	"encoding/json"
-	"fmt"
 	"io"
 	"io/ioutil"
 	"math/rand"
@@ -14,16 +12,13 @@ import (
 	"time"
 
 	"github.com/TRON-US/go-btfs/core"
-	"github.com/TRON-US/go-btfs/core/coreapi"
 	"github.com/TRON-US/go-btfs/core/coreunix"
 	"github.com/TRON-US/go-btfs/pin/gc"
 	"github.com/TRON-US/go-btfs/repo"
 
-	chunker "github.com/TRON-US/go-btfs-chunker"
 	config "github.com/TRON-US/go-btfs-config"
 	files "github.com/TRON-US/go-btfs-files"
 	coreiface "github.com/TRON-US/interface-go-btfs-core"
-	"github.com/TRON-US/interface-go-btfs-core/path"
 	blocks "github.com/ipfs/go-block-format"
 	"github.com/ipfs/go-blockservice"
 	cid "github.com/ipfs/go-cid"
@@ -33,8 +28,6 @@ import (
 	pi "github.com/ipfs/go-ipfs-posinfo"
 	dag "github.com/ipfs/go-merkledag"
 )
-
-const testPeerID = "QmTFauExutTsy4XP6JbMFcw2Wa9645HJt2bTqL6qYDCKfe"
 
 func TestAddMultipleGCLive(t *testing.T) {
 	r := &repo.Mock{
@@ -260,76 +253,7 @@ func TestAddGCLive(t *testing.T) {
 }
 
 func TestAddWithReedSolomonMetadata(t *testing.T) {
-	r := &repo.Mock{
-		C: config.Config{
-			Identity: config.Identity{
-				PeerID: testPeerID, // required by offline node
-			},
-		},
-		D: syncds.MutexWrap(datastore.NewMapDatastore()),
-	}
-	node, err := core.NewNode(context.Background(), &core.BuildCfg{Repo: r})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	out := make(chan interface{})
-	adder, err := coreunix.NewAdder(context.Background(), node.Pinning, node.Blockstore, node.DAG)
-	if err != nil {
-		t.Fatal(err)
-	}
-	adder.Out = out
-	// Set to default reed solomon for metadata
-	dsize, psize, csize := 10, 20, 262144
-	adder.Chunker = fmt.Sprintf("reed-solomon-%d-%d-%d", dsize, psize, csize)
-
-	fb := []byte("testfileA")
-	fsize := len(fb)
-	rfa := files.NewBytesFile(fb)
-
-	go func() {
-		defer close(out)
-		_, err := adder.AddAllAndPin(rfa)
-
-		if err != nil {
-			t.Fatal(err)
-		}
-	}()
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*5)
-	defer cancel()
-
-	var addedHash cid.Cid
-	select {
-	case o := <-out:
-		addedHash = o.(*coreiface.AddEvent).Path.Cid()
-	case <-ctx.Done():
-		t.Fatal("add timed out")
-	}
-
-	api, err := coreapi.NewCoreAPI(node)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// Extract and check metadata
-	b, err := coreunix.MetaData(ctx, api, path.IpfsPath(addedHash))
-	if err != nil {
-		t.Fatal(err)
-	}
-	var rsMeta chunker.RsMetaMap
-	err = json.Unmarshal(b, &rsMeta)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if rsMeta.NumData != uint64(dsize) {
-		t.Fatal("reed solomon metadata num data does not match")
-	}
-	if rsMeta.NumParity != uint64(psize) {
-		t.Fatal("reed solomon metadata num parity does not match")
-	}
-	if rsMeta.FileSize != uint64(fsize) {
-		t.Fatal("reed solomon metadata file size does not match")
-	}
+	HelpTestAddWithReedSolomonMetadata(t)
 }
 
 func testAddWPosInfo(t *testing.T, rawLeaves bool) {
